// src/genereted/api/defaultApi.ts
/**
 * Example API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { BirthExternal } from '../model/birthExternal.js';
import { Fruit } from '../model/fruit.js';
import { InlineObjects200Response } from '../model/inlineObjects200Response.js';
import { InlineObjectsRequest } from '../model/inlineObjectsRequest.js';
import { Profile } from '../model/profile.js';

export interface DefaultApi {
  dates: [{ birthday: string, bornAt: Date,  }, void];
  enums: [{  }, Array<Fruit>];
  hello: [{ name: string, expiresAt?: Date,  }, string];
  inlineObjects: [{ inlineObjectsRequest: InlineObjectsRequest,  }, InlineObjects200Response];
  multiFile: [{ file: RequestFile, files: Array<RequestFile>,  }, void];
  parameters: [{ id: string, limit?: number, q?: string, xRequestID?: string, sessionId?: string,  }, void];
  singleFile: [{ body: RequestFile,  }, void];
  withRefs: [{ profile: Profile,  }, BirthExternal];
}