// src/genereted/model/fruit.ts
/**
 * Example API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const Fruits = {
    Strawberry: 'STRAWBERRY',
    Cherry: 'CHERRY',
    Grape: 'GRAPE',
    Dekopon: 'DEKOPON',
    Orange: 'ORANGE',
    Apple: 'APPLE',
    Pear: 'PEAR',
    Peach: 'PEACH',
    Pineapple: 'PINEAPPLE',
    Melon: 'MELON',
    Watermelon: 'WATERMELON'
} as const;
export type Fruit = typeof Fruits[keyof typeof Fruits];