// src/genereted/model/birthExternal.ts
/**
 * Example API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { YearMonth } from './yearMonth.js';

export interface BirthExternal {
    'birthday'?: string;
    'born_at'?: Date;
    'birth_month'?: YearMonth;
}